# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dienstplan-App API für Persönliche Assistenz
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: Alle Benutzer auflisten
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, assistant]
      responses:
        "200":
          description: Liste der Benutzer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Neuen Benutzer anlegen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Erstellt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Benutzer abrufen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Benutzer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Nicht gefunden
    patch:
      operationId: updateUser
      tags: [users]
      summary: Benutzer aktualisieren
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Aktualisiert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Benutzer löschen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Gelöscht
...
  /time-tracking:
    get:
      operationId: listTimeEntries
      tags: [time-tracking]
      summary: Zeiteinträge auflisten
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, rejected]
      responses:
        "200":
          description: Liste der Zeiteinträge
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeEntry"
    post:
      operationId: createTimeEntry
      tags: [time-tracking]
      summary: Zeiteintrag erstellen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeEntryInput"
      responses:
        "201":
          description: Erstellt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
  /time-tracking/{id}:
    get:
      operationId: getTimeEntry
      tags: [time-tracking]
      summary: Zeiteintrag abrufen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Zeiteintrag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
        "404":
          description: Nicht gefunden
    patch:
      operationId: updateTimeEntry
      tags: [time-tracking]
      summary: Zeiteintrag aktualisieren
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeEntryUpdate"
      responses:
        "200":
          description: Aktualisiert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
    delete:
      operationId: deleteTimeEntry
      tags: [time-tracking]
      summary: Zeiteintrag löschen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Gelöscht
  /time-tracking/{id}/confirm:
    patch:
      operationId: confirmTimeEntry
      tags: [time-tracking]
      summary: Zeiteintrag bestätigen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeEntryConfirm"
      responses:
        "200":
          description: Bestätigt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, accountType, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, assistant]
        accountType:
          type: string
          enum: [privat, dienstleister]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        socialSecurityNumber:
          type: ["string", "null"]
        taxId:
          type: ["string", "null"]
        taxClass:
          type: ["string", "null"]
        healthInsurance:
          type: ["string", "null"]
        iban:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [name, email, role]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, assistant]
        phone:
          type: string
        address:
          type: string
        birthDate:
          type: string
        socialSecurityNumber:
          type: string
        taxId:
          type: string
        taxClass:
          type: string
        healthInsurance:
          type: string
        iban:
          type: string
        isActive:
          type: boolean
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        accountType:
          type: string
          enum: [privat, dienstleister]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        socialSecurityNumber:
          type: ["string", "null"]
        taxId:
          type: ["string", "null"]
        taxClass:
          type: ["string", "null"]
        healthInsurance:
          type: ["string", "null"]
        iban:
          type: ["string", "null"]
        isActive:
          type: boolean
...
    TimeEntry:
      type: object
      required: [id, userId, shiftId, actualStart, actualEnd, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        shiftId:
          type: ["integer", "null"]
        actualStart:
          type: string
          format: date-time
        actualEnd:
          type: string
          format: date-time
        actualHours:
          type: number
        status:
          type: string
          enum: [pending, confirmed, rejected]
        notes:
          type: ["string", "null"]
        confirmedBy:
          type: ["integer", "null"]
        confirmedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
    TimeEntryInput:
      type: object
      required: [userId, actualStart, actualEnd]
      properties:
        userId:
          type: integer
        shiftId:
          type: integer
        actualStart:
          type: string
          format: date-time
        actualEnd:
          type: string
          format: date-time
        notes:
          type: string
    TimeEntryUpdate:
      type: object
      properties:
        actualStart:
          type: string
          format: date-time
        actualEnd:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]
    TimeEntryConfirm:
      type: object
      required: [status, confirmedBy]
      properties:
        status:
          type: string
          enum: [confirmed, rejected]
        confirmedBy:
          type: integer
        notes:
          type: string
...