# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dienstplan-App API für Persönliche Assistenz
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateAccountType } from './userUpdateAccountType';

export interface UserUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  accountType?: UserUpdateAccountType;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  birthDate?: string | null;
  /** @nullable */
  socialSecurityNumber?: string | null;
  /** @nullable */
  taxId?: string | null;
  /** @nullable */
  taxClass?: string | null;
  /** @nullable */
  healthInsurance?: string | null;
  /** @nullable */
  iban?: string | null;
  isActive?: boolean;
}